change ducks file-structure
