+++ src/containers/MainAreaContainer.tsx
@@ -1,0 +1,3 @@
+import { MainArea } from '../components/MainArea'
+
+export const MainAreaConnected = MainArea
