+++ src/store/index.ts
@@ -1,0 +1,5 @@
+import { createStore } from 'redux'
+
+const store = createStore(() => ({}))
+
+export { store }
